requirements.txt
pdfplumber>=0.9.0
pandas>=1.5.0
openpyxl>=3.1.0
python-dateutil>=2.8.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0
click>=8.1.0
pyyaml>=6.0
pytest>=7.0.0
pytest-cov>=4.0.0
streamlit>=1.28.0
